pyproject.toml
[project]
name = "fastapi-setup"
version = "0.1.0"
description = ""
authors = [
    {name = "John Doe",email = "john@doe.com"}
]
requires-python = ">=3.13"

[tool.poetry.dependencies]
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
sqlmodel = "^0.0.24"
passlib = "^1.7.4"
bcrypt = "^3.2.2"
pydantic = {extras = ["email"], version = "^2.11.5"}
python-dotenv = "^1.1.0"
python-jose = "^3.4.0"
pydantic-settings = "^2.9.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
exclude = ["README.md"]